V3/Local_RL_Plant_System/requirements.txt
tensorflow
gymnasium
rich
opencv-python-headless
RPi.GPIO
adafruit-circuitpython-dht
spidev
PyYAML
numpy
